Optimize skybox drawing using early depth testing
Rendering the skybox before all objects with the depth test disabled
worked well but also meant that all visible fragments in the screen
space for the skybox were calculated every frame.
This is now optimized by setting the depth value for all skybox
fragments to the maximum value and rendering the skybox after all other
objects in the scene. This does have the added disavantage that skybox
is not visible through the window objects in the scene as the skybox
fragments are now depth tested and OpenGL determines that since the
window is closer to the camera and the skybox fragment being occluded
by the window can be discarded.