HDR: High Dynamic Range rendering
Use a seaparate floating point buffer to render a high dynamic range
image texture of the scene. This allows exceeding the [0, 1] range for
colors while drawing in the color buffer.
Use the color buffer from the new framebuffer to sample output pixels
and reduce color values to LDR by tonemapping and adjusting the exposure
factor to get the final render.